zad2/requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.4
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
httpie==0.9.9
idna==2.6
Markdown==2.6.11
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.3
Pygments==2.2.0
pytz==2018.4
requests==2.18.4
selenium==3.12.0
six==1.11.0
tox==3.0.0
urllib3==1.22
virtualenv==16.0.0
